WIP: support for weather underground

+++ src/wunder.h
@@ -1,22 +1,23 @@
 #ifndef _WUNDER_H
 #define _WUNDER_H
 
+#include <stdint.h>
 #include <time.h>
 
 struct ws_wunder
 {
-	time_t time;		/* timestamp */
+	time_t time;						/* timestamp */
 
-	int wind_dir;		/* wind direction */
-	double wind_speed;	/* wind speed (m/s) */
-	int humidity;		/* relative humidity */
-	double dew_point;	/* dew point (°C) */
-	double temp;		/* temperature (°C) */
-	double rain;		/* accumulated rainfall (mm) in the past 60 min */
-	double daily_rain;	/* accumulated rainfall (mm) today */
+	uint16_t wind_dir;					/* wind direction */
+	double wind_speed;					/* wind speed (m/s) */
+	uint8_t humidity;					/* relative humidity */
+	double dew_point;					/* dew point (°C) */
+	double temp;						/* temperature (°C) */
+	double rain;						/* accumulated rainfall (mm) in the past hour */
+	double daily_rain;					/* accumulated rainfall (mm) today */
 
-	double temp_in;		/* indoor temperature (°C) */
-	int humidity_in;	/* indoor humidity (°C) */
+	double temp_in;						/* indoor temperature (°C) */
+	uint8_t humidity_in;				/* indoor humidity (°C) */
 };
 
 int ws_wunder_upload(const struct ws_wunder *w);
